refactor(native): extract tauri setup
